Xamarin.Forms samples integration with samples folder
