+++ samples/Xamarin.Forms/references01project/XamarinAuth/Views/PageOAuthProviders.xaml.cs
@@ -36,112 +36,44 @@ namespace XamarinAuth.Views
 			switch (provider)
 			{
 				case "Facebook OAuth2":
-                    Xamarin.Auth.Helpers.OAuth2 oauth_facebook = null;
-                    oauth_facebook = Data.TestCases[provider] as Xamarin.Auth.Helpers.OAuth2;  
-                    page = new PageOAuth
-                                (
-                                    oauth_facebook.OAuth_IdApplication_IdAPI_KeyAPI_IdClient_IdCustomer,
-                                    oauth_facebook.OAuth2_Scope,
-                                    oauth_facebook.OAuth_UriAuthorization,
-                                    oauth_facebook.OAuth_UriCallbackAKARedirect,
-                                    null,
-                                    true
-                                );
+					Xamarin.Auth.Helpers.OAuth2 oauth_facebook = null;
+					oauth_facebook = Data.TestCases [provider] as Xamarin.Auth.Helpers.OAuth2;
+					AuthenticateOAuth2 (oauth_facebook);
 					break;
 				case "Twitter OAuth1":
                     Xamarin.Auth.Helpers.OAuth1 oauth_twitter = null;
                     oauth_twitter = Data.TestCases[provider] as Xamarin.Auth.Helpers.OAuth1;  
-                    page = new PageOAuth
-                                (
-                                    oauth_twitter.OAuth_IdApplication_IdAPI_KeyAPI_IdClient_IdCustomer,
-                                    oauth_twitter.OAuth1_SecretKey_ConsumerSecret_APISecret,
-                                    oauth_twitter.OAuth1_UriRequestToken,
-                                    oauth_twitter.OAuth_UriAuthorization,
-                                    oauth_twitter.OAuth_UriAccessToken,
-                                    oauth_twitter.OAuth_UriCallbackAKARedirect,
-                                    null,
-                                    true
-                                );
+					AuthenticateOAuth1 (oauth_twitter);
                 	break;
 				case "Google OAuth2":
                     Xamarin.Auth.Helpers.OAuth2 oauth_google = null;
                     oauth_google = Data.TestCases[provider] as Xamarin.Auth.Helpers.OAuth2;  
-                    page = new PageOAuth
-                                (
-                                    oauth_google.OAuth_IdApplication_IdAPI_KeyAPI_IdClient_IdCustomer,
-                                    oauth_google.OAuth2_Scope,
-                                    oauth_google.OAuth_UriAuthorization,
-                                    oauth_google.OAuth_UriCallbackAKARedirect,
-                                    null,
-                                    true
-                                );
+					AuthenticateOAuth2 (oauth_google);
                     break;
 				case "Microsoft Live OAuth2":
                     Xamarin.Auth.Helpers.OAuth2 oauth_mslive = null;
                     oauth_mslive = Data.TestCases[provider] as Xamarin.Auth.Helpers.OAuth2;  
-                    page = new PageOAuth
-                                (
-                                    oauth_mslive.OAuth_IdApplication_IdAPI_KeyAPI_IdClient_IdCustomer,
-                                    oauth_mslive.OAuth2_Scope,
-                                    oauth_mslive.OAuth_UriAuthorization,
-                                    oauth_mslive.OAuth_UriCallbackAKARedirect,
-                                    null,
-                                    true
-                                );
+					AuthenticateOAuth2 (oauth_mslive);
                     break;
 				case "LinkedIn OAuth1":
                     Xamarin.Auth.Helpers.OAuth1 oauth_linkedin1 = null;
                     oauth_linkedin1 = Data.TestCases[provider] as Xamarin.Auth.Helpers.OAuth1;  
-                    page = new PageOAuth
-                                (
-                                    oauth_linkedin1.OAuth_IdApplication_IdAPI_KeyAPI_IdClient_IdCustomer,
-                                    oauth_linkedin1.OAuth1_SecretKey_ConsumerSecret_APISecret,
-                                    oauth_linkedin1.OAuth1_UriRequestToken,
-                                    oauth_linkedin1.OAuth_UriAuthorization,
-                                    oauth_linkedin1.OAuth_UriAccessToken,
-                                    oauth_linkedin1.OAuth_UriCallbackAKARedirect,
-                                    null,
-                                    true
-                                );
+					AuthenticateOAuth1 (oauth_linkedin1);
                     break;
 				case "LinkedIn OAuth2":
                     Xamarin.Auth.Helpers.OAuth2 oauth_linkedin2 = null;
                     oauth_linkedin2 = Data.TestCases[provider] as Xamarin.Auth.Helpers.OAuth2;  
-                    page = new PageOAuth
-                                (
-                                    oauth_linkedin2.OAuth_IdApplication_IdAPI_KeyAPI_IdClient_IdCustomer,
-                                    oauth_linkedin2.OAuth2_Scope,
-                                    oauth_linkedin2.OAuth_UriAuthorization,
-                                    oauth_linkedin2.OAuth_UriCallbackAKARedirect,
-                                    null,
-                                    true
-                                );
+					AuthenticateOAuth2 (oauth_linkedin2);
                     break;
 				case "Github OAuth2":
                     Xamarin.Auth.Helpers.OAuth2 oauth_github = null;
                     oauth_github = Data.TestCases[provider] as Xamarin.Auth.Helpers.OAuth2;  
-                    page = new PageOAuth
-                                (
-                                    oauth_github.OAuth_IdApplication_IdAPI_KeyAPI_IdClient_IdCustomer,
-                                    oauth_github.OAuth2_Scope,
-                                    oauth_github.OAuth_UriAuthorization,
-                                    oauth_github.OAuth_UriCallbackAKARedirect,
-                                    null,
-                                    true
-                                );
+					AuthenticateOAuth2 (oauth_github);
                     break;
 				case "Instagram OAuth2":
                     Xamarin.Auth.Helpers.OAuth2 oauth_instagram = null;
                     oauth_instagram = Data.TestCases[provider] as Xamarin.Auth.Helpers.OAuth2;  
-                    page = new PageOAuth
-                                (
-                                    oauth_instagram.OAuth_IdApplication_IdAPI_KeyAPI_IdClient_IdCustomer,
-                                    oauth_instagram.OAuth2_Scope,
-                                    oauth_instagram.OAuth_UriAuthorization,
-                                    oauth_instagram.OAuth_UriCallbackAKARedirect,
-                                    null,
-                                    true
-                                );
+					AuthenticateOAuth2 (oauth_instagram);
                     break;
 				default:
 					DisplayAlert ("Error", "Unknown OAuth Provider!", "OK");
@@ -152,6 +84,42 @@ namespace XamarinAuth.Views
 
 			return;
 
+		}
+
+		public void AuthenticateOAuth2(Xamarin.Auth.Helpers.OAuth2 oauth2)
+		{
+			page = new PageOAuth
+				(
+					oauth2.OAuth_IdApplication_IdAPI_KeyAPI_IdClient_IdCustomer,
+					oauth2.OAuth2_Scope,
+					oauth2.OAuth_UriAuthorization,
+					oauth2.OAuth_UriCallbackAKARedirect,
+					null,
+					true
+				);
+
+			Navigation.PushAsync (page);
+
+			return;
+		}
+
+		public void AuthenticateOAuth1(Xamarin.Auth.Helpers.OAuth1 oauth1)
+		{
+			page = new PageOAuth
+				(
+					oauth1.OAuth_IdApplication_IdAPI_KeyAPI_IdClient_IdCustomer,
+					oauth1.OAuth1_SecretKey_ConsumerSecret_APISecret,
+					oauth1.OAuth1_UriRequestToken,
+					oauth1.OAuth_UriAuthorization,
+					oauth1.OAuth_UriAccessToken,
+					oauth1.OAuth_UriCallbackAKARedirect,
+					null,
+					true
+				);
+
+			Navigation.PushAsync (page);
+
+			return;
 		}
 
 		public async void Auth_Completed (object sender, AuthenticatorCompletedEventArgs ee)
