+++ source/XamarinForms/Xamarin.Auth.XamarinForms.iOS-Classic/packages.config
@@ -1,4 +1,4 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Xamarin.Forms" version="2.0.0.6490" targetFramework="monotouch1" />
+  <package id="Xamarin.Forms" version="2.0.1.6505" targetFramework="MonoTouch10" />
 </packages>
