+++ source/XamarinForms/Xamarin.Auth.XamarinForms/Views/PageOAuth.xaml
@@ -2,50 +2,8 @@
 <ContentPage 
 	xmlns="http://xamarin.com/schemas/2014/forms" 
 	xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml" 
-	x:Class="Xamarin.Auth.XamarinForms.Views.PageOAuth"
+	x:Class="Xamarin.Auth.XamarinForms.PageOAuth"
 	>
 	<ContentPage.Content>
 	</ContentPage.Content>
-
-	<!--
-	xmlns:local="clr-namespace:Xamarin.Auth.XamarinForms.Views"
-
-		<ContentPage.Padding>
-			<OnPlatform x:TypeArguments="Thickness">
-			<OnPlatform.iOS>
-			0, 20, 0, 0
-			</OnPlatform.iOS>
-			<OnPlatform.Android>
-			0, 0, 0, 0
-			</OnPlatform.Android>
-			<OnPlatform.WinPhone>
-			0, 0, 0, 0
-			</OnPlatform.WinPhone>
-			</OnPlatform>
-		</ContentPage.Padding>
-
-		<ContentPage.Content>
-		<ListView 
-			x:Name="listViewSamples"
-			ItemTapped="OnItemSelected"
-			ItemsSource="{x:Static local:Navigation.Samples}"
-			>
-			<ListView.ItemTemplate>
-				<DataTemplate>
-					<TextCell 
-						Text="{Binding Key}"
-						Detail="{Binding Value}"
-						>
-					</TextCell>
-				</DataTemplate>
-			</ListView.ItemTemplate>	
-		</ListView>
-
-	</ContentPage.Content>
--->
-<!--
-
-	XamarinForms/Xamarin.Auth.XamarinForms/Views/PageMain.xaml: 
-	Error: Target named 'UpdateDesignTimeXaml' not found in the project.
--->
 </ContentPage>
